fixing ssh and adding utils to prepare add signal data
