+++ webapp/controllers/api/Remote.js
@@ -26,53 +26,13 @@ module.exports = function(app) {
                   TcpManager.statusService(serviceInstance).then(function(result) {
                     console.log("Result: ", result);
                     // getting all data providers
-                    var dataProvidersResult = DataManager.listDataProviders();
-                    var providers = [];
-                    dataProvidersResult.forEach(function(dataProvider) {
-                      providers.push(dataProvider.toObject())
-                    }) // end foreach dataProvidersResult
-
-                    // getting dataseries
-                    DataManager.listDataSeries().then(function(dataSeriesResult) {
-                      var series = [];
-                      dataSeriesResult.forEach(function(dataSeries) {
-                        series.push(dataSeries.toObject());
-                      }) // end foreach dataSeriesResult
-
-                      // getting collectors
-                      DataManager.listCollectors().then(function(collectorsResult) {
-                        var collectors = [];
-                        collectorsResult.forEach(function(collector) {
-                          collectors.push(collector.toObject());
-                        }) // end foreach collectorsResult
-
-                        // getting analyses
-                        DataManager.listAnalyses().then(function(analysesResult) {
-                          var analyses = [];
-                          analysesResult.forEach(function(analysis) {
-                            analyses.push(analysis.toObject());
-                          }); // end foreach analysesResult
-
-                          // sending everything
-                          TcpManager.sendData(serviceInstance, {
-                            "Analysis": analyses,
-                            "DataSeries": series,
-                            "DataProviders": providers,
-                            "Collectors": collectors
-                          })
-
-                          // todo: check it/ping
-                          response.json({status: 200, online: Object.keys(result).length > 0});
-
-                        }).catch(function(err) {
-                          _handleError(response, err)
-                        }); // end listAnalyses
-                      }).catch(function(err) {
-                        _handleError(response, err)
-                      }); // end listCollectors
+                    Utils.prepareAddSignalMessage(DataManager).then(function(data) {
+                      TcpManager.sendData(serviceInstance, data);
+                      // todo: check it/ping
+                      response.json({status: 200, online: Object.keys(result).length > 0});
                     }).catch(function(err) {
                       _handleError(response, err)
-                    }); // end listDataSeries
+                    }); // end prepare
                   }).catch(function(err) {
                     _handleError(response, err);
                   }); // end statusService
