+++ webapp/core/Utils.js
@@ -5,6 +5,7 @@ var cloneDeep = require("lodash").cloneDeep;
 var crypto = require('crypto');
 var exceptions = require('./Exceptions');
 var Signals = require('./Signals');
+var Promise = require("bluebird");
 
 function getTokenCodeMessage(code) {
   var msg;
@@ -183,5 +184,52 @@ module.exports = {
 
   getUserHome: function() {
     return process.env[(process.platform == 'win32') ? 'USERPROFILE' : 'HOME'];
+  },
+
+  prepareAddSignalMessage: function(DataManager) {
+    return new Promise(function(resolve, reject) {
+      var _handleError = function(err) {
+        console.log(err);
+        reject(err);
+      }
+
+      var dataProvidersResult = DataManager.listDataProviders();
+      var providers = [];
+      dataProvidersResult.forEach(function(dataProvider) {
+        providers.push(dataProvider.toObject())
+      }) // end foreach dataProvidersResult
+
+      // getting dataseries
+      DataManager.listDataSeries().then(function(dataSeriesResult) {
+        var series = [];
+        dataSeriesResult.forEach(function(dataSeries) {
+          series.push(dataSeries.toObject());
+        }) // end foreach dataSeriesResult
+
+        // getting collectors
+        DataManager.listCollectors().then(function(collectorsResult) {
+          var collectors = [];
+          collectorsResult.forEach(function(collector) {
+            collectors.push(collector.toObject());
+          }) // end foreach collectorsResult
+
+          // getting analyses
+          DataManager.listAnalyses().then(function(analysesResult) {
+            var analyses = [];
+            analysesResult.forEach(function(analysis) {
+              analyses.push(analysis.toObject());
+            }); // end foreach analysesResult
+
+            resolve({
+              "Analysis": analyses,
+              "DataSeries": series,
+              "DataProviders": providers,
+              "Collectors": collectors
+            });
+
+          }).catch(_handleError); // end listAnalyses
+        }).catch(_handleError); // end listCollectors
+      }).catch(_handleError); // end listDataSeries
+    });
   }
 };
